fix(listenbrainz): address review findings
- Send name/mbid (not release_name/release_mbid) when logging a
  suggestion, and require a genre, so the insert actually succeeds;
  add a genre selector to the suggestion panel.
- Serialise syncs with a non-blocking lock and tolerate IntegrityError so
  a manual sync can't race the scheduler on the unique listen key.
- Truncate listen-weighted averages with int() so the no-data fallback
  matches the previous Bayesian scores exactly.
- Bound the "today" window by the next local midnight so DST days are
  23/25h rather than leaking an hour of the adjacent day.
- Retry 429s honouring ListenBrainz rate-limit headers and throttle
  listen fetches to one request per second.

+++ src/databass/db/models/catalog.py
@@ -1095,8 +1095,12 @@ class ArtistOrLabel(MusicBrainzEntity):
         if not rows:
             return []
 
+        # Truncate each entity average with int() exactly like
+        # average_ratings_bayesian(), so with no listen data the two produce
+        # identical scores (the weighted average is fractional once listens
+        # exist, but the display scale is whole 0-100 either way).
         weighted = [
-            (row.weighted_sum / row.total_weight, row.total_weight, row)
+            (int(row.weighted_sum / row.total_weight), row.total_weight, row)
             for row in rows
             if row.total_weight
         ]
